refactor(api): move remaining domain-logic helpers from lib to services
Continue the lib/services boundary cleanup per AGENTS.md placement rules.
Move the domain-logic helpers that remained in lib/ to services/:

- clipAndCreate.ts (URL→fetch→Readability→Tiptap→Y.Doc→DB pipeline)
- userDelete.ts (soft-delete business logic)
- auditLog.ts (admin audit-log domain helper)
- extAuth.ts (Chrome extension auth flow: Redis/PKCE/JWT)
- mcpAuth.ts (MCP auth flow: Redis/PKCE/JWT)
- freeEmailDomains.ts (domain-invitation validation logic)
- aiAccessHelpers.ts (AI model access checks; couples usage/aiProviders/subscription services)

Mirrored tests moved to src/__tests__/services/ and all imports followed.

src/lib/ now holds only generic helpers (clientIp, clipServerFetch,
clipUrlPolicy, cors, env, extractPlainTextFromYXml, githubAppAuth,
hocuspocusInvalidation, sentry, videoServerExtension,
youtubeEmbedServerExtension).

tsc --noEmit and full vitest suite (1425 tests) pass.

Refs #1013
