+++ server/api/src/routes/admin/index.ts
@@ -4,8 +4,8 @@ import { authRequired } from "../../middleware/auth.js";
 import { adminRequired } from "../../middleware/adminAuth.js";
 import { users, session } from "../../schema/users.js";
 import { pages } from "../../schema/pages.js";
-import { recordAuditLog } from "../../lib/auditLog.js";
-import { getUserImpact, anonymizeUser } from "../../lib/userDelete.js";
+import { recordAuditLog } from "../../services/auditLog.js";
+import { getUserImpact, anonymizeUser } from "../../services/userDelete.js";
 import auditLogsRoutes from "./auditLogs.js";
 import errorsRoutes from "./errors.js";
 import type { AppEnv } from "../../types/index.js";
