+++ server/api/src/services/clipAndCreate.ts
@@ -13,9 +13,9 @@ import { prosemirrorJSONToYDoc } from "@tiptap/y-tiptap";
 import type { NodePgDatabase } from "drizzle-orm/node-postgres";
 import { pages, pageContents } from "../schema/index.js";
 import type * as schema from "../schema/index.js";
-import { buildArticleSchema, extractArticleFromUrl } from "../services/articleExtractor.js";
+import { buildArticleSchema, extractArticleFromUrl } from "./articleExtractor.js";
 import type { AIProviderType, TokenUsage } from "../types/index.js";
-import { ensureDefaultNote } from "../services/defaultNoteService.js";
+import { ensureDefaultNote } from "./defaultNoteService.js";
 
 const YDOC_FRAGMENT = "default";
